Add overflow_counter
Extend a counter's count from an arbitrary bit width to 64-bits by
detecting overflows in the count. Each detected overflow is added to an
overflow counter which is combined with the current count in order create a
count up to 64-bits in length.

Resolves #63

+++ include/libembeddedhal/bit_limits.hpp
@@ -15,7 +15,7 @@ template<size_t BitWidth, std::integral T>
 static consteval T generate_field_of_ones()
 {
   T result = 0;
-  for (int i = 0; i < BitWidth; i++) {
+  for (size_t i = 0; i < BitWidth; i++) {
     result |= 1 << i;
   }
   return result;
